Proper double

+++ src/test/java/test/Base.java
@@ -30,8 +30,8 @@ import org.junit.internal.AssumptionViolatedException;
 class Base {
 
     protected void run() {
-        int failureOdd = Integer.valueOf(System.getProperty("failure.odd", "0.1"));
-        int skippingOdd = Integer.valueOf(System.getProperty("skipping.odd", "0.2"));
+        double failureOdd = Double.valueOf(System.getProperty("failure.odd", "0.1"));
+        double skippingOdd = Double.valueOf(System.getProperty("skipping.odd", "0.2"));
         double r = Math.random();
         if (r < failureOdd) {
             fail("oops");
